Refactor part state enums into separate files
