+++ service/SpaceCenter/src/KRPC.SpaceCenter.csproj
@@ -97,21 +97,29 @@
     <Compile Include="Services\Orbit.cs" />
     <Compile Include="Services\Parts\Decoupler.cs" />
     <Compile Include="Services\Parts\DockingPort.cs" />
+    <Compile Include="Services\Parts\DockingPortState.cs" />
     <Compile Include="Services\Parts\Engine.cs" />
     <Compile Include="Services\Parts\LandingGear.cs" />
+    <Compile Include="Services\Parts\LandingGearState.cs" />
     <Compile Include="Services\Parts\LandingLeg.cs" />
+    <Compile Include="Services\Parts\LandingLegState.cs" />
     <Compile Include="Services\Parts\LaunchClamp.cs" />
     <Compile Include="Services\Parts\Light.cs" />
     <Compile Include="Services\Parts\Module.cs" />
     <Compile Include="Services\Parts\Parachute.cs" />
+    <Compile Include="Services\Parts\ParachuteState.cs" />
     <Compile Include="Services\Parts\Part.cs" />
     <Compile Include="Services\Parts\Parts.cs" />
     <Compile Include="Services\Parts\Radiator.cs" />
+    <Compile Include="Services\Parts\RadiatorState.cs" />
     <Compile Include="Services\Parts\ReactionWheel.cs" />
     <Compile Include="Services\Parts\ResourceConverter.cs" />
+    <Compile Include="Services\Parts\ResourceConverterState.cs" />
     <Compile Include="Services\Parts\ResourceHarvester.cs" />
+    <Compile Include="Services\Parts\ResourceHarvesterState.cs" />
     <Compile Include="Services\Parts\Sensor.cs" />
     <Compile Include="Services\Parts\SolarPanel.cs" />
+    <Compile Include="Services\Parts\SolarPanelState.cs" />
     <Compile Include="Services\ReferenceFrame.cs" />
     <Compile Include="Services\Resources.cs" />
     <Compile Include="Services\SASMode.cs" />
