+++ service/SpaceCenter/src/Services/Parts/LandingGearState.cs
@@ -1,0 +1,28 @@
+using KRPC.Service.Attributes;
+
+namespace KRPC.SpaceCenter.Services.Parts
+{
+    /// <summary>
+    /// See <see cref="LandingGear.State"/>.
+    /// </summary>
+    [KRPCEnum (Service = "SpaceCenter")]
+    public enum LandingGearState
+    {
+        /// <summary>
+        /// Landing gear is fully deployed.
+        /// </summary>
+        Deployed,
+        /// <summary>
+        /// Landing gear is fully retracted.
+        /// </summary>
+        Retracted,
+        /// <summary>
+        /// Landing gear is being deployed.
+        /// </summary>
+        Deploying,
+        /// <summary>
+        /// Landing gear is being retracted.
+        /// </summary>
+        Retracting
+    }
+}
