+++ service/SpaceCenter/src/Services/Parts/LandingLeg.cs
@@ -6,38 +6,6 @@ using KRPC.SpaceCenter.ExtensionMethods;
 namespace KRPC.SpaceCenter.Services.Parts
 {
     /// <summary>
-    /// See <see cref="LandingLeg.State"/>.
-    /// </summary>
-    [KRPCEnum (Service = "SpaceCenter")]
-    public enum LandingLegState
-    {
-        /// <summary>
-        /// Landing leg is fully deployed.
-        /// </summary>
-        Deployed,
-        /// <summary>
-        /// Landing leg is fully retracted.
-        /// </summary>
-        Retracted,
-        /// <summary>
-        /// Landing leg is being deployed.
-        /// </summary>
-        Deploying,
-        /// <summary>
-        /// Landing leg is being retracted.
-        /// </summary>
-        Retracting,
-        /// <summary>
-        /// Landing leg is broken.
-        /// </summary>
-        Broken,
-        /// <summary>
-        /// Landing leg is being repaired.
-        /// </summary>
-        Repairing
-    }
-
-    /// <summary>
     /// Obtained by calling <see cref="Part.LandingLeg"/>.
     /// </summary>
     [KRPCClass (Service = "SpaceCenter")]
