+++ service/SpaceCenter/src/Services/Parts/Parachute.cs
@@ -6,35 +6,6 @@ using KRPC.SpaceCenter.ExtensionMethods;
 namespace KRPC.SpaceCenter.Services.Parts
 {
     /// <summary>
-    /// See <see cref="Parachute.State"/>.
-    /// </summary>
-    [KRPCEnum (Service = "SpaceCenter")]
-    public enum ParachuteState
-    {
-        /// <summary>
-        /// The parachute is still stowed, but ready to semi-deploy.
-        /// </summary>
-        Active,
-        /// <summary>
-        /// The parachute has been cut.
-        /// </summary>
-        Cut,
-        /// <summary>
-        /// The parachute is fully deployed.
-        /// </summary>
-        Deployed,
-        /// <summary>
-        /// The parachute has been deployed and is providing some drag,
-        /// but is not fully deployed yet.
-        /// </summary>
-        SemiDeployed,
-        /// <summary>
-        /// The parachute is safely tucked away inside its housing.
-        /// </summary>
-        Stowed
-    }
-
-    /// <summary>
     /// Obtained by calling <see cref="Part.Parachute"/>.
     /// </summary>
     [KRPCClass (Service = "SpaceCenter")]
