+++ service/SpaceCenter/src/Services/Parts/Radiator.cs
@@ -6,34 +6,6 @@ using KRPC.SpaceCenter.ExtensionMethods;
 namespace KRPC.SpaceCenter.Services.Parts
 {
     /// <summary>
-    /// <see cref="RadiatorState"/>
-    /// </summary>
-    [KRPCEnum (Service = "SpaceCenter")]
-    public enum RadiatorState
-    {
-        /// <summary>
-        /// Radiator is fully extended.
-        /// </summary>
-        Extended,
-        /// <summary>
-        /// Radiator is fully retracted.
-        /// </summary>
-        Retracted,
-        /// <summary>
-        /// Radiator is being extended.
-        /// </summary>
-        Extending,
-        /// <summary>
-        /// Radiator is being retracted.
-        /// </summary>
-        Retracting,
-        /// <summary>
-        /// Radiator is being broken.
-        /// </summary>
-        Broken
-    }
-
-    /// <summary>
     /// Obtained by calling <see cref="Part.Radiator"/>.
     /// </summary>
     [KRPCClass (Service = "SpaceCenter")]
